+++ client/settings.ini
@@ -1,0 +1,4 @@
+[settings]
+REMOTE_URL=http://citz-imb-vrm-route-e1c611-dev.apps.silver.devops.gov.bc.ca:5555
+;REMOTE_URL=http://localhost:5555
+CLIENT_ENCODING=utf-8
