+++ server/static/js/script.js
@@ -1,0 +1,7 @@
+document.getElementById('uploadBtn').addEventListener('click', function() {
+    document.getElementById('fileInput').click();
+});
+
+document.getElementById('fileInput').addEventListener('change', function() {
+    this.form.submit();
+});
